+++ VivaDictaKeyboard/KeyboardViewController.swift
@@ -130,55 +130,7 @@ struct ActivateButton: View {
     }
 }
 
-struct MicButton: View {
-    @State var isAnimating = false
-    
-    var fontSize: CGFloat
-    var padding: CGFloat
-    var backgroundColor: Color
-    var borderWidth: CGFloat
-    
-    var onTapAction: () -> Void
-    
-    
-    var body: some View {
-        
-        
-        Button {
-            onTapAction()
-        } label: {
-            
-            Image(systemName: "microphone.circle")
-                .foregroundColor(.primary)
-                .font(.system(size: fontSize))
-                .padding(padding)
-                .background(backgroundColor.gradient, in: .circle)
-                
-                .background {
-                    Circle()
-                    
-                        .fill(AngularGradient(colors: [.teal, .pink, .teal], center: .center, angle: .degrees(isAnimating ? 360 : 0)))
-                        .blur(radius: 10)
-                        .onAppear {
-                            withAnimation(Animation.linear(duration: 7).repeatForever(autoreverses: false)) {
-                                isAnimating = true
-                            }
-                        }
-                        .onDisappear {
-                            isAnimating = false
-                        }
-                }
-            
-                .overlay {
-                    Circle()
-                        .stroke(.black.opacity(0.5), lineWidth: borderWidth)
-                }
-        }
-        
-        
-        
-    }
-}
+
 
 
 struct VivaDictaKeyboardToolbarView: View {
